backapp/requirements.txt
fastapi
uvicorn
python-dotenv
requests
duckduckgo-search
google-generativeai
Pillow
torch
torchvision
numpy
scikit-learn
google-search-results
